Conflicts:
	.gitignore
	core/QC.sh
	core/Summary.py
	core/makeSummary.sh

+++ mods/fseq.sh
@@ -5,9 +5,9 @@
 # date: January 1914
 
 # QCVARIABLES,Resource temporarily unavailable
-# RESULTFILENAME <DIR>/<TASK>/<SAMPLE>.bw
+# RESULTFILENAME <DIR>/<TASK>/<SAMPLE>.narrowPeak
 
-echo ">>>>> DNase-Seq with fseq"
+echo ">>>>> Peak calling with fseq"
 echo ">>>>> startdate "`date`
 echo ">>>>> hostname "`hostname`
 echo ">>>>> job_name "$JOB_NAME
@@ -42,9 +42,10 @@ done
 ################################################################################
 CHECKPOINT="programs"
 
-for MODULE in $MODULE_FSEQ; do module load $MODULE; done  # save way to load modules that itself load other modules
+# save way to load modules that itself loads other modules
+hash module 2>/dev/null && for MODULE in $MODULE_FSEQ; do module load $MODULE; done && module list 
+
 export PATH=$PATH_FSEQ:$PATH
-module list
 echo "PATH=$PATH"
 
 echo "[NOTE] set java parameters"
@@ -60,7 +61,11 @@ echo -e "--bedtools --\n "$(bedtools --version)
 echo -e "--wigToBigWig --\n "$(wigToBigWig 2>&1 | tee | head -n 1)
 [ -z "$(which wigToBigWig)" ] && echo "[ERROR] wigToBigWig not detected" && exit 1
 echo -e "--fseq        --\n "$(fseq -v | head -n 1)
-[ -z "$(which | head -n 1)" ] && echo "[ERROR] no | head -n 1 detected" && exit 1
+[ -z "$(which fseq)" ] && echo "[ERROR] no fseq detected" && exit 1
+echo -e "--R           --\n "$(R --version | head -n 3)
+[ -z "$(which R)" ] && echo "[ERROR] no R detected" && exit 1
+echo -e "--bedToBigBed --\n "$(bedToBigBed 2>&1 | tee | head -n 1 )
+[ -z "$(which bedToBigBed)" ] && echo "[WARN] bedToBigBed not detected, cannot compress bedgraphs"
 
 echo -e "\n********* $CHECKPOINT\n"
 ################################################################################
@@ -69,11 +74,11 @@ CHECKPOINT="parameters"
 # get basename of input file f
 INPUTFILENAME=${INPUTFILE##*/}
 # get sample prefix
-SAMPLE=${INPUTFILENAME/%$bam/}
+SAMPLE=${INPUTFILENAME/%$ASD.bam/}
 
 # delete old bam files unless attempting to recover
 if [ -z "$RECOVERFROM" ]; then
-    ## TODO remove primary result files from pervious runs
+    [ -f $OUTDIR/$SAMPLE.bw ] && rm $OUTDIR/$SAMPLE.bw
 fi
 
 if [ -z "$FASTA" ] || [ ! -f $FASTA ]; then
@@ -91,13 +96,12 @@ else
     echo "[NOTE] Chromosome size: $GENOME_CHROMSIZES"
 fi
 
-
-if [ -n "$FSEQ_PLOIDY" ]; then
-    FSEQ_PLOIDY_PARAM="-p $FSEQ_PLOIDY"
+if [ -z "$FSEQ_PVALUECUTOFF" ] || [ $FSEQ_PVALUECUTOFF -gt 1 ] ; then
+    FSEQ_PVALUECUTOFF=0.05
 fi
 
 # unique temp folder that should be used to store temporary files
-THISTMP=$TMP"/"$(whoami)"/"$(echo $OUTDIR | md5sum | cut -d' ' -f1)
+THISTMP=$TMP"/"$(whoami)"/"$(echo $OUTDIR/$SAMPLE | md5sum | cut -d' ' -f1)
 mkdir -p $THISTMP
 
 echo -e "\n********* $CHECKPOINT\n"
@@ -113,25 +117,89 @@ fi
 echo -e "\n********* $CHECKPOINT\n"
 
 ################################################################################
-CHECKPOINT="fseq"
+CHECKPOINT="fseq bw"
 
 if [[ -n "$RECOVERFROM" ]] && [[ $(grep -P "^\*{9} $CHECKPOINT" $RECOVERFROM | wc -l ) -gt 0 ]] ; then
     echo "::::::::: passed $CHECKPOINT"
 else
 
-    rm -f $THISTMP/*wig
-
-#    RUN_COMMAND="bedtools bamtobed -i $f | fseq $FSEQADDPARAM $FSEQ_PLOIDY_PARAM -b $FSEQ_BACKGROUNDDIR -o $THISTMP -of wig "
-    RUN_COMMAND="bedtools bamtobed -i $f | java $JAVAOPTS -cp $CLASSPATH edu.duke.igsp.gkde.Main $FSEQADDPARAM $FSEQ_PLOIDY_PARAM -b $FSEQ_BACKGROUNDDIR -o $THISTMP -of wig "
-
-    echo $RUN_COMMAND && eval $RUN_COMMAND
+    if [ -n "$FSEQ_MAKEBIGWIG" ]; then
+        rm -f $THISTMP/*wig
     
-    RUN_COMMAND="cat $THISTMP/*.wig | wigToBigWig stdin ${GENOME_CHROMSIZES} $OUTDIR/SAMPLE.bw"
+        RUN_COMMAND="java $JAVAPARAMS -cp $CLASSPATH edu.duke.igsp.gkde.Main $FSEQADDPARAM -o $THISTMP -of wig <(bedtools bamtobed -i $INPUTFILE )"
+        echo $RUN_COMMAND && eval $RUN_COMMAND
+        
+        cat $THISTMP/*.wig | wigToBigWig stdin ${GENOME_CHROMSIZES} $OUTDIR/$SAMPLE.bw
+    
+        # mark checkpoint
+        if [ -f $OUTDIR/$SAMPLE.bw ];then echo -e "\n********* $CHECKPOINT\n"; unset RECOVERFROM; else echo "[ERROR] checkpoint failed: $CHECKPOINT"; exit 1; fi
+    else
+        echo "[NOTE] skip bw generation"
+        echo -e "\n********* $CHECKPOINT\n"
+    fi 
+fi
+################################################################################
+CHECKPOINT="fseq narrowpeak"
+
+if [[ -n "$RECOVERFROM" ]] && [[ $(grep -P "^\*{9} $CHECKPOINT" $RECOVERFROM | wc -l ) -gt 0 ]] ; then
+    echo "::::::::: passed $CHECKPOINT"
+else
+
+    RUN_COMMAND="java $JAVAPARAMS -cp $CLASSPATH edu.duke.igsp.gkde.Main $FSEQADDPARAM -o $THISTMP -of npf <(bedtools bamtobed -i $INPUTFILE )"
     echo $RUN_COMMAND && eval $RUN_COMMAND
+
+    cat $THISTMP/*.npf > $OUTDIR/$SAMPLE.narrowPeak
 
     # mark checkpoint
-    if [ -f $OUTDIR/$SAMPLE.bw ];then echo -e "\n********* $CHECKPOINT\n"; unset RECOVERFROM; else echo "[ERROR] checkpoint failed: $CHECKPOINT"; exit 1; fi
+    if [ -f $OUTDIR/$SAMPLE.narrowPeak ];then echo -e "\n********* $CHECKPOINT\n"; unset RECOVERFROM; else echo "[ERROR] checkpoint failed: $CHECKPOINT"; exit 1; fi
 fi
+
+################################################################################
+CHECKPOINT="estimate p-values"
+
+if [[ -n "$RECOVERFROM" ]] && [[ $(grep -P "^\*{9} $CHECKPOINT" $RECOVERFROM | wc -l ) -gt 0 ]] ; then
+    echo "::::::::: passed $CHECKPOINT"
+else
+
+    # estimate p-value by fitting gamma distribution
+    cat > $OUTDIR/$SAMPLE.R <<DELIM    
+library(MASS)
+x <- read.delim("$OUTDIR/$SAMPLE.narrowPeak", header=F)
+g<-MASS::fitdistr(x\$V7,"gamma")
+x\$V8<-sapply(x\$V7, function(y) ks.test(y, "pgamma",shape=g\$estimate["shape"],rate=g\$estimate["rate"])\$p.value)
+write.table(x, "$OUTDIR/$SAMPLE.narrowPeak.tmp", sep="\t", row.names=FALSE, col.names=FALSE, quote=FALSE)
+DELIM
+    
+    Rscript --vanilla $OUTDIR/$SAMPLE.R
+
+    #filter with p-value cutoff
+    awk -v CUTOFF=$FSEQ_PVALUECUTOFF '{if ($8 <= CUTOFF){print $0}}' $OUTDIR/$SAMPLE.narrowPeak.tmp > $OUTDIR/$SAMPLE.narrowPeak
+    [ -f $OUTDIR/$SAMPLE.narrowPeak.tmp ] && rm $OUTDIR/$SAMPLE.narrowPeak.tmp 
+    [ -f $OUTDIR/$SAMPLE.R ] && rm $OUTDIR/$SAMPLE.R
+
+    # mark checkpoint
+    if [ -f $OUTDIR/$SAMPLE.narrowPeak ];then echo -e "\n********* $CHECKPOINT\n"; unset RECOVERFROM; else echo "[ERROR] checkpoint failed: $CHECKPOINT"; exit 1; fi
+fi
+
+################################################################################
+CHECKPOINT="generate bigbed"
+
+if [[ -n "$RECOVERFROM" ]] && [[ $(grep -P "^\*{9} $CHECKPOINT" $RECOVERFROM | wc -l ) -gt 0 ]] ; then
+    echo "::::::::: passed $CHECKPOINT"
+else
+    # convert to bigbed
+	if hash bedToBigBed ; then 
+        echo "[NOTE] create bigbed from peaks" 
+        awk '{OFS="\t"; print $1,$2,$3,$7}' $OUTDIR/$SAMPLE.narrowPeak > $OUTDIR/$SAMPLE.peak.tmp
+        bedToBigBed -type=bed4 $OUTDIR/$SAMPLE.peak.tmp $GENOME_CHROMSIZES $OUTDIR/$SAMPLE.bb
+        rm $OUTDIR/$SAMPLE.peak.tmp
+         # mark checkpoint
+        if [ -f $OUTDIR/$SAMPLE.bb ];then echo -e "\n********* $CHECKPOINT\n"; unset RECOVERFROM; else echo "[ERROR] checkpoint failed: $CHECKPOINT"; exit 1; fi
+    else
+        echo "[NOTE] bigbed not generated"
+        echo -e "\n********* $CHECKPOINT\n"
+    fi
+fi   
 ###############################################################################
 CHECKPOINT="cleanup"
 
@@ -139,6 +207,6 @@ CHECKPOINT="cleanup"
 
 echo -e "\n********* $CHECKPOINT\n"
 ################################################################################
-[ -e $OUTDIR/$SAMPLE.bw.dummy ] && rm $OUTDIR/$SAMPLE.bw.dummy
-echo ">>>>> DNase-Seq with fseq - FINISHED"
+[ -e $OUTDIR/$SAMPLE.narrowPeak.dummy ] && rm $OUTDIR/$SAMPLE.narrowPeak.dummy
+echo ">>>>> Peak calling with fseq - FINISHED"
 echo ">>>>> enddate "`date`
